pyproject.toml
[tool.poetry]
name = "otus-onnx-model-serving"
version = "0.1.0"
description = ""
authors = ["akhmadiev <akhmadiev@tax-monitor.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.135.3"}
uvicorn = {extras = ["standard"], version = "^0.44.0"}
onnxruntime = "^1.24.4"
numpy = "^2.4.4"
python-multipart = "^0.0.24"
pydantic = "^2.12.5"
python-dotenv = "^1.2.2"
pyjwt = "^2.12.1"
psutil = "^7.2.2"
bcrypt = "<4.0.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
config = "^0.5.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
